feat(scenario): set is_gamemaster when creating a post
This feature is tricky, because the first turn in a thread
has {charactersCount} posts, while every turns after the first
have {charactersCount + 1} to account for the GameMaster.

We have to keep track of the current turn and the previous posters
to determine if a new post is a GameMaster's one.