feat(ux): improve Characters list on home page
If it's this character turn to play, the UI will now display a small warning.
I'm not really proud of the way I achieved that on the backend side,
but I'm tired and it's been a long week at work.
"But captain, it's only wednesday!"
Yes, but my week started a month ago actually.

Anyway, if it's dumb but it works, it's not so dumb after all.